Refactor icons and logo images (#67)
