+++ frontend/src/components/pages/home/header/navigation/index.tsx
@@ -1,6 +1,6 @@
 import styled, { useTheme } from 'styled-components';
 import NavigationItem from './NavigationItem';
-import { ReactComponent as Menu } from '../../../../../icons/menu.svg';
+import { ReactComponent as Menu } from '../../../../../img/icons/menu.svg';
 import useWindowDimensions from '../../../../../util/windowDimensions';
 import { useSetRecoilState } from 'recoil';
 import { sidebarOpenState, SidebarType, sidebarTypeState } from '../../../../../state/home/sidebar';
