+++ frontend/src/components/pages/panel/ProductListItem.tsx
@@ -2,8 +2,8 @@ import { Draggable } from 'react-beautiful-dnd';
 import styled, { useTheme } from 'styled-components';
 import { Product } from '../../../api/types';
 import Typography from '../../common/Typography';
-import { ReactComponent as Delete } from '../../../icons/delete.svg';
-import { ReactComponent as Edit } from '../../../icons/edit.svg';
+import { ReactComponent as Delete } from '../../../img/icons/delete.svg';
+import { ReactComponent as Edit } from '../../../img/icons/edit.svg';
 import { useState } from 'react';
 import { useSetRecoilState } from 'recoil';
 import { popupProductImageState } from '../../../state/panel/products';
